@@ -23,22 +23,22 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  provider: ^3.0.0
-  firebase_core: 0.7.0
-  firebase_auth: ^0.20.0+1
-  cloud_firestore: ^0.16.0
-  firebase_storage: ^7.0.0
-  firebase_crashlytics: ^0.4.0+1
+  provider: ^5.0.0
+  firebase_core: ^1.0.1
+  firebase_auth: ^1.0.1
+  cloud_firestore: ^1.0.1
+  firebase_storage: ^8.0.0
+  firebase_crashlytics: ^1.0.0
   firebase_analytics: ^7.0.1
-  webview_flutter: ^0.3.18+1
+  webview_flutter: ^2.0.2
   image_picker:
   autocomplete_textfield: ^1.7.3
   slugify: ^1.0.0
   circular_menu: ^1.1.2
   cached_network_image: ^2.5.0
-  google_sign_in: ^4.5.1
+  google_sign_in: ^5.0.0
   tap_debouncer: ^1.0.1
-  auth_buttons: ^0.0.6
+  auth_buttons: ^1.0.0
   fluttertoast: ^7.1.8
 
   # The following adds the Cupertino Icons font to your application.
